services/requirements.txt
requests==2.28.1
kafka-python==2.0.2
beautifulsoup4==4.11.1
